Added the ability to instantiate transport plugins within the emulator (internal transports) as an NEM layer stack entry. This eliminates the need for running emanetransportd in all use cases where emane and emanetransportd are running on the same node.
Transports that are internal to the NEM no longer require
platformendpoint and transportendpoint configuration parameters. In
order to connect an emulator with an external transport, either running
in emanetransportd or embedded in another application, you will have to
add a transport='external' attribute to the <nem> platform XML elements
and specify both platformendpoint and transportendpoint. Internal
transports are now the default.

In order to support internal transports, DTDs were changed to remove the
often confusing <transport> is not in the proper place when defining an
NEM. Existing XML will have to place <transport> at the top of layer
definitions instead of the bottom, this is where it logically belongs.

The Remote Control Port protobuf specification and emanesh were modified
to support configuration, statistic and statistic table access for
internal transports.

emanegentransportxml was modified to ignore internal transports.

Fixed flow control token grant acknowledgment bug. See
https://github.com/adjacentlink/emane/issues/8.