Do not use '-g0' in CXXFLAGS
If CXXFLAGS in the environment contains '-g', do not override it.
rpmbuild and dpkg-buildpackage purposefully add this flag to create
a separate package containing debugging symbols, which are stripped
from the binaries after they are built. The debugging symbols are
very useful to trace the behavior of EMANE with utilities such as
SystemTap, and this includes optimized builds.

Signed-off-by: David Ward <david.ward@ll.mit.edu>