Added flow control functionality. Flow control requires a transport capable of supporting flow control, such as the virtual transport.
Modified the QueueManager API to return the number of packets dropped
(if any) during an enqueue due to overflow. Dropped packet information
is necessary for increasing available tokens.

Added an additional flow control outbound drop code for reporting
bytes dropped due to flow control. Drops due to flow control are an
indication of an issue with token synchronization and should not
occur.

Trailing white space cleanup.