Added additional SchedulerUser functionality to support processSchedulerPacket(), processSchedulerControl() and getPacketQueueInfo(). SchedulerUser processSchedulerPacket() method was modified to remove the control messages parameter from call to address potential ambiguity due to aggregation and fragmentation functions. The non-packet control message path, processSchedulerControl() should be used by a scheduler module wishing to send control messages to the physical layer.
Queue functionality and accessor method was added to track the number
of packets and bytes in a queue. Partial packets remaining in the
queue due to fragmentation are counted as packets in the queue, but
only the bytes remaining are counted towards bytes in the queue.

Queue initialization was modified to add an indication as to queue
type: data or control. This information is used when creating message
components in order to route control messages to a receive-side
scheduler module.

Trailing white space cleanup.