Added aggregation enable information to low level queues in order to disable in-queue search for additional message components when aggregation is disabled.
Modified queue fragmentation disabled drop policy. Packets too large
for the slot will be discarded whiling searching for one that
fits. Once one packet is found, aggregation will be abandoned, if there
is more room in the slot but the next packet in the queue is too
large. Previous logic would have continued to drop packets while
searching for one that fit.