emane.spec: Use automatic dependency processing for external libraries
Do not specify external libraries with the Requires tag; rpmbuild
adds them during automatic dependency processing (except for those
needed to run scripts). The packages specified by the Requires tag
are not sufficient as-is. Also the automatic dependencies have the
advantage of using the library SONAME rather than the package name
to ensure that installed library versions are ABI-compatible with
the versions that were linked with EMANE.

Signed-off-by: David Ward <david.ward@ll.mit.edu>