Use constexpr specifier for endianness conversion functions
This permits values such as EMANE::HTONS(Utils::ETH_P_IPV4) to be
computed at compile-time, which is already possible when using the
equivalent preprocessor definitions in glibc. This is required in
order to use such value in enums, as an example. Remove the inline
specifier, since constexpr functions are implicitly inline.

Signed-off-by: David Ward <david.ward@ll.mit.edu>