Add missing get() wrapper function in DownstreamPacket class
An EMANE::DownstreamPacket::get() function was available prior to
version 0.9.x, which returned a pointer to the packet data (first
combining packet segments if necessary). It has been moved to the
EMANE::DownstreamPacket::Implementation class, but there is not a
wrapper function to call it from EMANE::DownstreamPacket. This is
useful in situations where the MAC layer does not simply prepend
headers to downstream packets, but consumes them and creates its
own downstream packets instead.

Signed-off-by: David Ward <david.ward@ll.mit.edu>