Under high traffic load and certain datarate configurations, when downstream traffic arrives it is possible for the packet queue to be empty but the previous packet end-of-transmission (EOT) time to still be in the future. Failure to check for that condition leads to immediate transmission of the newly arrived packet instead of waiting for EOT.
Additional code was added to processDownstreamPacket() to test for a
currentEndOfTransmissionTime_ in the future in order to determine if a
callback should be scheduled or if the newly arrived packet can be
immediately sent.

Additional code was added to handleDownstreamQueueEntry() to engage
transmission logic even if the scheduled handler had been short
circuited as long as there exists a packet to transmit in the queue
and currentEndOfTransmissionTime_ has past.

Configurable jitter and fixed delay logic was modified to apply the
values when figuring out currentEndOfTransmissionTime_ for the
purposes of determining the earliest possible time another packet may
be transmitted.

See https://github.com/adjacentlink/emane/issues/40