Removed remaining ACE framework code and replaced with C++11 or native linux equivalent.
Removed code for platforms unsupported since 0.9: WIN32 and OSX.

Support for emulator use of the system logger was removed.

Support for gpsd location agent direct connection with gpsd was
removed. This feature worked inconsistently and was not
recommended. Instead, gpsd should be configured to use the pseudo
terminal name as the gps device.

Removed the PROBLEM-REPORT-FORM. It was outdated and rarely used.

White space cleanup.