An issue occurred when using floating point representation with K, M or G SI unit values for integer configuration items when the result after applying the appropriate power of 10 contained one or more leading 0s.
A leading 0 caused the result to be treated as an octal value. If the
parameter value only contained values in the set of allowable octal
characters, the conversion would succeed but as base 8. If the value
contained digits outside the octal set (e.g. 0.9G), conversion would
fail with a not allowed character exception description.

Functionality was added to remove any leading zeros after SI unit
application minding the possibility of sign inclusion and/or leading
white space. This effectively results in all integer values with SI
units being interpreted as base 10.

Reported-by: bebopagogo <https://github.com/bebopagogo>

See https://github.com/adjacentlink/emane/issues/45