Remove duration and time point type parameters in function templates
In the EMANE::Utils::Timer class, the duration and time point types
used for the member variables and callback parameters are fixed. As a
result, there does not appear to be a benefit to parameterizing these
types for the function templates schedule() and scheduleInterval().

However, these function templates perform type casts that can require
additional computation and cause a loss of precision. To illustrate
this, consider when they are instantiated using an EMANE::TimePoint
for the absoluteTimePoint parameter, which has nanosecond resolution
(with libstdc++). The function templates both type-cast the parameter
to have microsecond resolution, but then they immediately pass this
as an argument to the (non-template) function schedule(), where it is
implicitly type-cast back to an EMANE::TimePoint again.

Avoid this by declaring schedule() and scheduleInterval() to have the
same fixed types for duration and time point as the (non-template)
function schedule(). Callers of these function templates should make
use of C++11 implicit type-casting for durations and time points.

Signed-off-by: David Ward <david.ward@ll.mit.edu>