Introduce EMANE::Duration type alias, and use it for timer intervals
The timer interval is specified as a duration in microseconds, but
instead it should have the same resolution as the expiration time.
One reason is that these two values are added together frequently.
Since they both describe emulation time, it would also make sense
for them to have the same resolution.

Expiration time is given as an EMANE::TimePoint, a type alias for
EMANE::Clock::time_point. In libstdc++, time points for all clock
types are implemented with nanosecond resolution. EMANE::Duration
is now a type alias for EMANE::Clock::duration and should be used
when specifying timer intervals.

Existing models will continue to compile without any changes. When
calling EMANE::TimerService*::scheduleTimedEvent() and specifying
a timer interval, type-casting it from microseconds to nanoseconds
is implicit in C++11.

Signed-off-by: David Ward <david.ward@ll.mit.edu>