configure: Set default value for CXXFLAGS as described in Autoconf manual
If the CXXFLAGS environment variable is not set when ./configure is
run, then CXXFLAGS will be initialized to an Autoconf default value
of "-g -O2", as part of the AC_PROG_CXX macro. This macro can use a
different default value instead, as described in the Autoconf manual
(section 5.10.4). Use that approach to specify our default value for
CXXFLAGS, instead of appending default options to CXXFLAGS after the
macro has run. This will allow our default value for CXXFLAGS to be
suppressed if desired, by setting the CXXFLAGS environment variable.

When building rpm or deb packages, the CXXFLAGS environment variable
is generally set with a distribution-supplied value. Ensure that our
default options in CXXFLAGS are still used by appending them to this
environment variable, using the RPM SPEC file or debian/rules file.

Note that the -g option is retained from the Autoconf default value
for CXXFLAGS. If debug symbols are not desired after building from
the source code directly, the user should run 'make install-strip'
instead of 'make install' according to the Automake manual (section
2.2.2). This behavior is unchanged from before.

Signed-off-by: David Ward <david.ward@ll.mit.edu>