configure: Improve options for CPPFLAGS, CXXFLAGS, and LDFLAGS
The options for debug symbols (-g), profiling information (-pg), and
optimization (-O) are used by the compiler and not the preprocessor,
so place them in CXXFLAGS instead of CPPFLAGS.

When --with-debug is passed to ./configure, add the -Og option rather
than -O0, if it is supported (gcc 4.8 or later). Include additional
macros from Autoconf Archive to check for this support.

Signed-off-by: David Ward <david.ward@ll.mit.edu>