configure: Fix output variables for date and distribution name
The RELDATE variable is still used when building man pages, but it
is not being set by the configure script. The man pages expect the
date in ISO 8601 format.

To obtain the distribution name, set the DISTRIB_CODENAME variable
by sourcing /etc/lsb-release. (Do not parse /etc/debian_version as
this provides the wrong value.) Use this variable in the changelog
for .deb packages, replacing the DEBIAN_VERSION variable.

Signed-off-by: David Ward <david.ward@ll.mit.edu>