Makefile: Improve rpm target
Do not pass the --target option to rpmbuild (and do not set the RPMARCH
variable in the Makefile). rpmbuild expects the --target option to have
the format CPU-VENDOR-OS which is not currently being given. However the
target does not need to be specified anyway, as the current architecture
is used automatically. (Building for a different target architecture is
better achieved using a build environment such as mock.)

Setting DISTCLEANFILES to remove .rpmbuild is unneeded, since it is
removed by the clean-local target.

Signed-off-by: David Ward <david.ward@ll.mit.edu>