Makefile: List input files in appropriate variables
The protocol buffer compiler (protoc) is required to build EMANE and
is used to generate C++ code from protocol definition files (.proto)
during the build process. The generated code is not included in the
source distributions (emane-X.Y.Z.tar.gz), so protoc should not be
triggered by "make dist". Use the "nodist_" prefix to list generated
C++ code files that should be used in a program or library but not
included in the source distribution, as described in the Automake
documentation:
https://www.gnu.org/software/automake/manual/html_node/Program-and-Library-Variables.html

The Automake documentation recommends listing header files that will
not be installed in the system (i.e. in /usr/include) in a SOURCES
variable (so move these files from the EXTRA_DIST variable):
https://www.gnu.org/software/automake/manual/html_node/Headers.html

The above changes make it easier to list protoc-generated files in
BUILT_SOURCES by referencing the existing nodist_*_SOURCES variable,
and to keep the Makefile shorter. This also eliminates the need for
the dist-hook target in the top-level Makefile.

Signed-off-by: David Ward <david.ward@ll.mit.edu>