Rename "emane-python" package to "python-emane" and improve packaging
Fedora and Debian both provide packaging guidelines for Python modules:

https://www.debian.org/doc/packaging-manuals/python-policy/
https://fedoraproject.org/wiki/Packaging:Python

Following these guidelines, use the package name "python-emane" since it
(will eventually) contain a Python module named "emane". Make the Debian
package architecture-independent with a section name of "python". In the
Fedora and Red Hat packages use the %{python_sitelib} macro when listing
the installed files for the Python module.

Specify Python Setuptools as a build dependency for this package. Change
the runtime dependency in the Debian package to python-pkg-resources.

Add the missing license file to the Fedora and Red Hat packages.

Signed-off-by: David Ward <david.ward@ll.mit.edu>