Improve package dependencies and fix other package tags
Add dependencies in the Debian/Ubuntu packages that resemble those in
the Fedora/Red Hat packages. A package for a layer or another plugin
now depends on the application using the plugin, and a package for an
application now depends on the DTD files.

Allow transport plugins to be installed if either the platform daemon
(emane) or transport daemon (emanetransportd) is installed, since the
platform daemon can instantiate transport layers since EMANE 0.9.x.

Use architecture-specific and fully-versioned dependencies in Fedora/
Red Hat between subpackages, following the recommendations here:
https://fedoraproject.org/wiki/Packaging:Guidelines

Mark the packages that contain DTD files, manifest files, and scripts
as architecture-independent.

Replace the Debian section names "extra" and "binary" as they are not
valid. Following the Debian-provided package for ns3, use the section
name "net" for the application packages (implicitly) and the source
package (explicitly).

Do not make packages for Fedora and Red Hat relocatable (by using the
Prefix tag). Attempting to relocate packages during their installation
is likely to cause unexpected behavior.

Signed-off-by: David Ward <david.ward@ll.mit.edu>