Makefile: Fix linking issues and missing symbol warnings
When building packages for Debian/Ubuntu, debhelper prints warnings
about missing symbols in the dh_shlibdeps stage. The changes below
fix these warnings by properly linking all programs and libraries
as instructed in the documentation for Automake and libtool.

Link against libemane and/or libemanexmlparser in every program and
library that uses its symbols. The Automake documentation explains
that a program should link against such a libtool library, which is
built from the same package and not yet installed in the system, by
listing the path to the library's .la file in the LDADD variable of
the program:
https://www.gnu.org/software/automake/manual/html_node/Libtool-Libraries.html

(This will ultimately link the program against the .so file for the
library, as desired.) This method is recommended over using the -l
and -L options to link against a library from the same package:
https://www.gnu.org/software/automake/manual/html_node/Linking.html

Likewise a library should be linked against another libtool library
by listing the .la file in its LIBADD variable (rather than LDADD):
https://www.gnu.org/software/automake/manual/html_node/Libtool-Flags.html

To link against external libraries that are already installed, the
linker flags obtained during ./configure (e.g. libxml2_LIBS) should
be added to the LDADD or LIBADD variable respectively. The LDFLAGS
variable is intended to be used to pass other flags to the linker
such as "-avoid-version" (see the second reference above).

The paths to header files for both internal and external libraries
are given to the preprocessor, using the -I flag, in the CPPFLAGS
variable. These flags for external libraries are obtained during
./configure as well (e.g. libxml2_CFLAGS). The CXXFLAGS variable
is instead specifically for options to the C++ compiler itself.

Signed-off-by: David Ward <david.ward@ll.mit.edu>