Makefile: Fix install-exec-hook and clean-local targets for emanesh
Pass the "--prefix" and "--exec-prefix" options used with ./configure
to "python setup.py install", so that the default prefix, /usr/local,
is overridden during packaging. Adjust the file list for the Debian
package to reflect this fix.

Do not pass the --root option to "python setup.py install" unless
$(DESTDIR) is set, to avoid installation underneath the source tree.
This should prevent the generation of .pyc files inside the source
tree as a result.

Remove setup.py during "make clean". Use the BUILT_SOURCES variable
to avoid redundancy.

Signed-off-by: David Ward <david.ward@ll.mit.edu>