Added BaseModel::Implementation::slotPortionRatio method to simplify slot portion ratio calculations used when issuing slot status table updates.
Updated SlotStatusTablePublisher update logic to handle input change
from the portion of the slot remaining to the portion of the slot
used.

Fixed slot portion ratio calculation for RX_MISSED. Previous logic was
not using the slot time associated with the start of transmission slot
and instead was using the actual received slot time. This lead to
erroneous quantile counts of less than 1.0 for all RX_MISSED packets.

White space cleanup.