Add frameworkphy to public interface
This allows less code duplication for users wanting to customize the
frameworkphy.

Also fixed an inconsistency with the nembuilder.cc.  Previously, the
SpectrumMonitor object is passed to the built-in frameworkphy PHY layer
and not custom PHY layers.  This effectively renders the
SpectrumServiceProvider useless for MAC layers sitting "above" a custom
PHY.  This commit passes the SpectrumMonitor object to all PHY layers.
This has a side-benefit of simplifying the layerfactorymanager.