Fix the physical layer checkHorizon calculation. checkHorizon calculates the distance to the earth's horizon as a function of a NEM's altitude. The calculation is performed to determine whether two NEMs are mutually in line of sight above the curved surface of the earth. The physical layer drops packets between two nodes when their summed distance to horizon is less than their distance apart.
The distance to horizion (d) is calculated as:

    For h <= 0:

      d = 0

      and

    For h > 0:

      d^2 = (R + h)^2 - R^2
          = 2Rh + h^2

          or

        d = sqrt(2Rh + h^2)

where:
     R is the mean earth radius
     h is the NEM altitude above the earth's surface

The existing calculation assumed h << R and computed d as:

    d = sqrt(2Rh)

Ignoring the h^2 term underestimates d. For cases where a NEM altitude is >= 2R,
the error is large enough for checkHorizon to drop all packets to/from any other
NEM near the surface of the earth.

The new calculation eliminates the error by retaining the h^2 term.