Added Doppler shift processing to the emulator physical layer. Doppler shift is calculated when both the transmitter and receiver location and velocity are known at the receiver, and the physical layer 'dopplershiftenable' parameter is set to 'on'. Setting 'dopplershiftenable' to 'off' disables Doppler shift processing.
Doppler shift is applied when energy is added to spectrum monitor bins
based on bandwidth overlap. Additionally, Doppler shift information is
communicated to compatibility mode 2 radio models via the
MIMOReceivePropertiesControlMessage:

  const DopplerShifts &
    MIMOReceivePropertiesControlMessage::getDopplerShifts() const;

Where, DopplerShifts is a std::map with a frequency in Hz
(frequencyofinterest) key and shift value in Hz.

Updated the physical layer ReceivePowerTable to add a Doppler column
containing the last computed Doppler shift in Hz.