Added radio silence support to the physical layer. Radio silence is controlled entirely in the physical layer and its current enable state is not communicated to the radio model. Packets dropped when radio silence is enabled are dropped before any other processing.
Radio silence is controlled by the phyisical layer configuration
parameter 'radiosilenceenable'. When 'radiosilenceenable' is set to
'on' all transmissions are dropped. The parameter is running-state
modifiable, and defaults to 'off'.

Added a new statistic numDownstreamPacketsRadioSilenceEnabledDrop
which indicates the total packets dropped/not transmitted due to
'radiosilenceenable' being set to 'on'